Refactor search page, migrate to /discover/
