+++ apps/frontend/src/locales/en-US/index.json
@@ -968,6 +968,9 @@
   "dashboard.withdraw.error.tax-form.title": {
     "message": "Please complete tax form"
   },
+  "discover.title": {
+    "message": "Discover"
+  },
   "error.collection.404.list_item.1": {
     "message": "You may have mistyped the collection's URL."
   },
