+++ apps/frontend/src/pages/collection/[id].vue
@@ -364,7 +364,7 @@
 					<span v-if="auth.user && auth.user.id === creator.id" class="preserve-lines text">
 						<IntlFormatted :message-id="messages.noProjectsAuthLabel">
 							<template #create-link="{ children }">
-								<a class="link" @click.prevent="$router.push('/mods')">
+								<a class="link" @click.prevent="$router.push('/discover/mods')">
 									<component :is="() => children" />
 								</a>
 							</template>
