+++ apps/frontend/src/pages/discover/index.vue
@@ -1,0 +1,6 @@
+<script setup lang="ts">
+throw createError({
+	fatal: true,
+	statusCode: 404,
+})
+</script>
